requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.0
pytesseract>=0.3.10
Pillow>=9.0.0
pandas>=1.5.0
langchain>=0.0.300
openai>=0.28.0
duckduckgo-search>=3.9.0 